reshuffling and adding tests
